refactor(watch): package layout + broaden integ checks
- constants.py: DEFAULT_MAX_PENDING, DEFAULT_POLL_INTERVAL, DIR_DETECTOR
  (both defaults are ours, not from #450; documented as such)
- queue/ package: WatchQueue protocol + OverflowPolicy + errors in
  queue/base.py, RAMWatchQueue in queue/ram.py
- source.py: Source/Subscriber runtime dataclasses out of watcher.py
- drop the default_queue helper; RAMWatchQueue is itself a valid
  QueueFactory (root-only ctor), so it is the default directly
- keep the change model (ChangeKind/ResourceChange/Delta) in types.py as
  a shared leaf: the producer (Workspace.watch) and the watch machinery
  both need it, so a neutral leaf avoids a workspace<->watch cycle
  without a TYPE_CHECKING workaround
- integ: nextcloud.json cases now assert cat/head/ls/grep after
  create/update/delete, proving reads are fresh post-invalidation (grep
  no longer matches a deleted file; old content gone after update)
