+++ integ/watch/nextcloud.json
@@ -7,27 +7,42 @@
   "cases": [
     {
       "id": "create_file",
-      "mutate": {"op": "write", "path": "data/new.txt", "body": "hello"},
-      "expect": {"kind": "create", "path": "/nc/data/new.txt"},
-      "verify": {"cmd": "cat /nc/data/new.txt", "contains": "hello"}
+      "mutate": {"op": "write", "path": "data/report.txt", "body": "alpha line\nbeta line\n"},
+      "expect": {"kind": "create", "path": "/nc/data/report.txt"},
+      "checks": [
+        {"cmd": "cat /nc/data/report.txt", "contains": "alpha line"},
+        {"cmd": "head -n 1 /nc/data/report.txt", "contains": "alpha line"},
+        {"cmd": "ls /nc/data", "contains": "report.txt"},
+        {"cmd": "grep -rl beta /nc/data", "contains": "report.txt"}
+      ]
     },
     {
       "id": "update_file",
-      "mutate": {"op": "write", "path": "data/new.txt", "body": "hello-v2"},
-      "expect": {"kind": "update", "path": "/nc/data/new.txt"},
-      "verify": {"cmd": "cat /nc/data/new.txt", "contains": "hello-v2"}
+      "mutate": {"op": "write", "path": "data/report.txt", "body": "gamma line\ndelta line\n"},
+      "expect": {"kind": "update", "path": "/nc/data/report.txt"},
+      "checks": [
+        {"cmd": "cat /nc/data/report.txt", "contains": "gamma line"},
+        {"cmd": "grep -rl gamma /nc/data", "contains": "report.txt"},
+        {"cmd": "grep -r alpha /nc/data", "absent": "report.txt"}
+      ]
     },
     {
       "id": "create_nested",
-      "mutate": {"op": "write", "path": "data/sub/deep.txt", "body": "deep"},
+      "mutate": {"op": "write", "path": "data/sub/deep.txt", "body": "deep marker\n"},
       "expect": {"kind": "create", "path": "/nc/data/sub/deep.txt"},
-      "verify": {"cmd": "cat /nc/data/sub/deep.txt", "contains": "deep"}
+      "checks": [
+        {"cmd": "cat /nc/data/sub/deep.txt", "contains": "deep marker"},
+        {"cmd": "grep -rl 'deep marker' /nc/data", "contains": "deep.txt"}
+      ]
     },
     {
       "id": "delete_file",
-      "mutate": {"op": "delete", "path": "data/new.txt"},
-      "expect": {"kind": "delete", "path": "/nc/data/new.txt"},
-      "verify": {"cmd": "ls /nc/data", "absent": "new.txt"}
+      "mutate": {"op": "delete", "path": "data/report.txt"},
+      "expect": {"kind": "delete", "path": "/nc/data/report.txt"},
+      "checks": [
+        {"cmd": "ls /nc/data", "absent": "report.txt"},
+        {"cmd": "grep -rl gamma /nc/data", "absent": "report.txt"}
+      ]
     }
   ]
 }
